// types/database.extended.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export type Measurement = {
  amount: number;
  unitLong: string;
  unitShort: string;
};

export type Measurements = {
  metric: Measurement;
  us: Measurement;
};

export type Database = {
  graphql_public: {
    Tables: {
      [_ in never]: never;
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      graphql: {
        Args: {
          operationName?: string;
          query?: string;
          variables?: Json;
          extensions?: Json;
        };
        Returns: Json;
      };
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
  public: {
    Tables: {
      ingredients: {
        Row: {
          id: number;
          measurements: Measurements | null;
          name: string | null;
          original: string | null;
          original_name: string | null;
          recipe_id: number;
        };
        Insert: {
          id?: number;
          measurements?: Measurements | null;
          name?: string | null;
          original?: string | null;
          original_name?: string | null;
          recipe_id: number;
        };
        Update: {
          id?: number;
          measurements?: Measurements | null;
          name?: string | null;
          original?: string | null;
          original_name?: string | null;
          recipe_id?: number;
        };
        Relationships: [
          {
            foreignKeyName: 'ingredients_recipe_id_fkey';
            columns: ['recipe_id'];
            isOneToOne: false;
            referencedRelation: 'recipes';
            referencedColumns: ['id'];
          }
        ];
      };
      recipes: {
        Row: {
          added_by: string | null;
          author: string | null;
          created_at: string;
          description: string | null;
          id: number;
          image_url: string | null;
          ingredients: string[] | null;
          notes: string[] | null;
          nutrition: Json | null;
          original_url: string | null;
          preparation: string[] | null;
          rating: number | null;
          servings: number | null;
          title: string | null;
        };
        Insert: {
          added_by?: string | null;
          author?: string | null;
          created_at?: string;
          description?: string | null;
          id?: number;
          image_url?: string | null;
          ingredients?: string[] | null;
          notes?: string[] | null;
          nutrition?: Json | null;
          original_url?: string | null;
          preparation?: string[] | null;
          rating?: number | null;
          servings?: number | null;
          title?: string | null;
        };
        Update: {
          added_by?: string | null;
          author?: string | null;
          created_at?: string;
          description?: string | null;
          id?: number;
          image_url?: string | null;
          ingredients?: string[] | null;
          notes?: string[] | null;
          nutrition?: Json | null;
          original_url?: string | null;
          preparation?: string[] | null;
          rating?: number | null;
          servings?: number | null;
          title?: string | null;
        };
        Relationships: [];
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type PublicSchema = Database[Extract<keyof Database, 'public'>];

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema['Tables'] & PublicSchema['Views'])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions['schema']]['Tables'] &
        Database[PublicTableNameOrOptions['schema']]['Views'])
    : never = never
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions['schema']]['Tables'] &
      Database[PublicTableNameOrOptions['schema']]['Views'])[TableName] extends {
      Row: infer R;
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema['Tables'] &
      PublicSchema['Views'])
  ? (PublicSchema['Tables'] &
      PublicSchema['Views'])[PublicTableNameOrOptions] extends {
      Row: infer R;
    }
    ? R
    : never
  : never;
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema['Enums']
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions['schema']]['Enums']
    : never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions['schema']]['Enums'][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema['Enums']
  ? PublicSchema['Enums'][PublicEnumNameOrOptions]
  : never;

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof PublicSchema['CompositeTypes']
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database;
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions['schema']]['CompositeTypes']
    : never = never
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions['schema']]['CompositeTypes'][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof PublicSchema['CompositeTypes']
  ? PublicSchema['CompositeTypes'][PublicCompositeTypeNameOrOptions]
  : never;